+++ CHANGELOG.md
@@ -8,6 +8,8 @@ title: Changelog
 
 -   `@include` and `@includeCode` now work in the readme file, #2814.
 -   TypeDoc will now avoid making references to references, #2811.
+-   Improved link resolution logic to prioritize type alias properties with the
+    same symbol over type literal properties within function parameters.
 
 ## v0.27.5 (2024-12-14)